# docs/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tavanir AI Assistant V2",
    "description": "Enterprise RAG & Suggestion Committee Assistant Service (MOCK MODE)",
    "version": "2.0.0"
  },
  "paths": {
    "/api/v1/suggestions/analyze": {
      "post": {
        "tags": [
          "v1",
          "Suggestions"
        ],
        "summary": "Analyze an incoming employee suggestion against historical suggestions",
        "description": "Synchronously executes Tri-Track hybrid Qdrant retrieval, candidate deduplication, cross-encoder reranking, Max-Passage pooling, and PostgreSQL hydration to return partitioned similar suggestion IDs.",
        "operationId": "analyze_suggestion_api_v1_suggestions_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeSuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_AnalyzeSuggestionDataResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Malformed syntax, invalid parameters, or payload violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API Key in header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Target resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource already exists or concurrent edit collision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or domain business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/suggestions/ingest": {
      "post": {
        "tags": [
          "v1",
          "Suggestions"
        ],
        "summary": "Ingest a historical or newly submitted employee suggestion",
        "description": "Synchronously normalizes, chunks, embeds, and stores the suggestion in PostgreSQL and Qdrant vector store.",
        "operationId": "ingest_suggestion_api_v1_suggestions_ingest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestSuggestionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_IngestSuggestionDataResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Malformed syntax, invalid parameters, or payload violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API Key in header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Target resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource already exists or concurrent edit collision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or domain business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/api/v1/suggestions/{suggestionId}": {
      "put": {
        "tags": [
          "v1",
          "Suggestions"
        ],
        "summary": "Full replacement update for an employee suggestion",
        "description": "Synchronously replaces all core fields of an existing suggestion or restores a soft-deleted suggestion with two-phase concurrency control.",
        "operationId": "update_suggestion_api_v1_suggestions__suggestionId__put",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "suggestionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Unique suggestion identifier",
              "title": "Suggestionid"
            },
            "description": "Unique suggestion identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSuggestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_UpdateSuggestionDataResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Malformed syntax, invalid parameters, or payload violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API Key in header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Target resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource already exists or concurrent edit collision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or domain business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "v1",
          "Suggestions"
        ],
        "summary": "Partial update for an employee suggestion",
        "description": "Synchronously overlays non-null fields onto an active suggestion with two-phase concurrency control. Soft-deleted suggestions cannot be patched.",
        "operationId": "patch_suggestion_api_v1_suggestions__suggestionId__patch",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "suggestionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Unique suggestion identifier",
              "title": "Suggestionid"
            },
            "description": "Unique suggestion identifier"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PatchSuggestionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_UpdateSuggestionDataResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Malformed syntax, invalid parameters, or payload violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API Key in header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Target resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource already exists or concurrent edit collision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or domain business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "v1",
          "Suggestions"
        ],
        "summary": "Soft delete a suggestion and purge its vector chunks",
        "description": "Guards against race conditions via advisory locking, sets is_deleted=True in PostgreSQL, and physically purges vector chunks from Qdrant.",
        "operationId": "delete_suggestion_api_v1_suggestions__suggestionId__delete",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "suggestionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "description": "Unique suggestion identifier",
              "title": "Suggestionid"
            },
            "description": "Unique suggestion identifier"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_DeleteSuggestionDataResponse_"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request - Malformed syntax, invalid parameters, or payload violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API Key in header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Target resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource already exists or concurrent edit collision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or domain business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/suggestions/bulk-delete": {
      "post": {
        "tags": [
          "v1",
          "Suggestions"
        ],
        "summary": "Bulk soft-delete suggestions",
        "description": "Sequentially soft-deletes up to 100 suggestions with per-item error isolation. Returns 200 OK if all succeed, 207 Multi-Status on partial success, or 400 Bad Request if all fail.",
        "operationId": "bulk_delete_suggestions_api_v1_suggestions_bulk_delete_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkDeleteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "All requested suggestions were successfully soft-deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse_list_DeleteSuggestionDataResponse__"
                }
              }
            }
          },
          "400": {
            "description": "All requested items failed to delete or input list violated constraints.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid API Key in header",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Target resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Resource already exists or concurrent edit collision",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation error or domain business rule violation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "207": {
            "description": "Partial success: some suggestions were deleted while others failed.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PartialSuccessResponse_DeleteSuggestionDataResponse_"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AnalyzeSuggestionRequest": {
        "properties": {
          "title": {
            "type": "string",
            "minLength": 5,
            "title": "Title",
            "description": "Title of the suggestion under evaluation"
          },
          "currentProblem": {
            "type": "string",
            "minLength": 5,
            "title": "Currentproblem",
            "description": "Description of the organizational problem or gap"
          },
          "solution": {
            "type": "string",
            "minLength": 5,
            "title": "Solution",
            "description": "Proposed engineering, operational, or software method"
          },
          "contextTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexttitle",
            "description": "Organizational department or technical context"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "title",
          "currentProblem",
          "solution"
        ],
        "title": "AnalyzeSuggestionRequest",
        "description": "Inbound request schema for suggestion analysis (ADR-002, 03_Naming_And_Data_Exchange.md).\nAccepts camelCase fields from API clients while binding to internal Python fields."
      },
...
      "ErrorResponse": {
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ErrorItem"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of errors associated with the request failure."
          }
        },
        "type": "object",
        "required": [
          "errors"
        ],
        "title": "ErrorResponse",
        "description": "The canonical top-level JSON:API error envelope."
      },
...
      "IngestSuggestionRequest": {
        "properties": {
          "suggestionId": {
            "type": "string",
            "minLength": 1,
            "title": "Suggestionid",
            "description": "Unique business identifier of the suggestion"
          },
          "title": {
            "type": "string",
            "minLength": 1,
            "title": "Title",
            "description": "Title of the suggestion"
          },
          "problem": {
            "type": "string",
            "minLength": 1,
            "title": "Problem",
            "description": "Description of the organizational problem/gap"
          },
          "solution": {
            "type": "string",
            "minLength": 1,
            "title": "Solution",
            "description": "Proposed solution or improvement"
          },
          "status": {
            "title": "Status",
            "description": "Evaluation status (Persian string e.g. 'مصوب', ID int, or enum name)"
          },
          "committeeScrutiny": {
            "title": "Committeescrutiny",
            "description": "Expert committee review/scrutiny (Persian title, code int, or string)"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Final resolution or execution commentary"
          },
          "shamsiDate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Shamsidate",
            "description": "Submission or evaluation date in Shamsi format (YYYY/MM/DD)"
          },
          "contextTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Contexttitle",
            "description": "Organizational department or domain context"
          },
          "secretariatScrutiny": {
            "title": "Secretariatscrutiny",
            "description": "Secretariat primary scrutiny (Persian title, code int, or string)"
          },
          "secretariatComment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secretariatcomment",
            "description": "Secretariat review rationale or commentary"
          },
          "tributaryScrutiny": {
            "title": "Tributaryscrutiny",
            "description": "Legacy alias for secretariat scrutiny"
          },
          "tributaryComment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tributarycomment",
            "description": "Legacy alias for secretariat commentary"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "suggestionId",
          "title",
          "problem",
          "solution",
          "status"
        ],
        "title": "IngestSuggestionRequest",
        "description": "Inbound request schema for single suggestion ingestion.\nAccepts camelCase fields from API clients while binding to snake_case internally\nvia BaseRequestModel's alias_generator=to_camel and populate_by_name=True."
      },
...
      "SuccessResponse_AnalyzeSuggestionDataResponse_": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code.",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/AnalyzeSuggestionDataResponse",
            "description": "Response payload data."
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[AnalyzeSuggestionDataResponse]"
      },
...
      "SuccessResponse_IngestSuggestionDataResponse_": {
        "properties": {
          "status": {
            "type": "integer",
            "title": "Status",
            "description": "HTTP status code.",
            "default": 200
          },
          "data": {
            "$ref": "#/components/schemas/IngestSuggestionDataResponse",
            "description": "Response payload data."
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "SuccessResponse[IngestSuggestionDataResponse]"
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}